requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-stubs==0.12.1
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mypy==0.701
mypy-extensions==0.4.1
Pygments==2.4.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
typed-ast==1.3.5
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.25.2
